Rewrite all MicroSamples as "real" XM apps
- All of the MicroSamples were badly broken (were even using mcs)
- Good example of "command line" tool in XM
