+++ MicroSamples/.gitignore
@@ -1,5 +1,1 @@
-demo.pdf
 *.app/
-HelloCoreWlanSample
-generate-pdf
-locator
