+++ MicroSamples/CoreWlan/CoreWlan.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreWlan", "CoreWlan.csproj", "{FFD1C590-12B1-4154-AB63-C18F2C45FDAF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FFD1C590-12B1-4154-AB63-C18F2C45FDAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFD1C590-12B1-4154-AB63-C18F2C45FDAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFD1C590-12B1-4154-AB63-C18F2C45FDAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFD1C590-12B1-4154-AB63-C18F2C45FDAF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
