+++ MicroSamples/locator/locator.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "locator", "locator.csproj", "{929B23F9-7A6C-4C1A-802B-D58562EADC63}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{929B23F9-7A6C-4C1A-802B-D58562EADC63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{929B23F9-7A6C-4C1A-802B-D58562EADC63}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{929B23F9-7A6C-4C1A-802B-D58562EADC63}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{929B23F9-7A6C-4C1A-802B-D58562EADC63}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
