+++ MicroSamples/pdf/pdf.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pdf", "pdf.csproj", "{6CC099BE-D6FC-4A03-AD32-6836B6DC7CD7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6CC099BE-D6FC-4A03-AD32-6836B6DC7CD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CC099BE-D6FC-4A03-AD32-6836B6DC7CD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CC099BE-D6FC-4A03-AD32-6836B6DC7CD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CC099BE-D6FC-4A03-AD32-6836B6DC7CD7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
